+++ routes/app.js
@@ -1,0 +1,55 @@
+import 'dotenv/config';
+
+import webpack from 'webpack';
+import webpackMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
+import config from '../webpack/webpack.config';
+
+const env = process.env.NODE_ENV || 'development';
+
+const routes = (app) => {
+
+  if (env === 'development') { // eslint-disable-line eqeqeq
+    const compiler = webpack(config);
+    const middleware = webpackMiddleware(compiler, {
+      publicPath: config.output.publicPath,
+      contentBase: 'src',
+      stats: {
+        colors: true,
+        hash: false,
+        timings: true,
+        chunks: false,
+        chunkModules: false,
+        modules: false,
+      },
+    });
+
+    app.use(middleware);
+    app.use(webpackHotMiddleware(compiler));
+
+    app.get('/*', (req, res) => {
+      const key = '<div id="root"></div>';
+      const content = middleware.fileSystem.readFileSync(path.join(__dirname, 'dist/index.html'));
+      const index = content.indexOf(key) + key.length;
+      const inject = req.user ? `<script type="text/javascript">window.id="${req.user._id}";</script>` : '';
+      res.send(content.slice(0, index) + inject + content.slice(index));
+    });
+
+  } else {
+
+    app.use(express.static(`${__dirname}`));
+
+    const content = fs.readFileSync(path.join(__dirname, 'dist/index.html'), 'utf8');
+    const key = '<div id="root"></div>';
+    const index = content.indexOf(key) + key.length;
+
+    app.get('/*', (req, res) => {
+      const inject = req.user ? `<script type="text/javascript">window.id="${req.user._id}";</script>` : '';
+      res.send(content.slice(0, index) + inject + content.slice(index));
+    });
+
+  }
+
+};
+
+export default routes;
